closures

+++ closures/src/lib.rs
@@ -1,3 +1,3 @@
 fn first_fifty_even_square() -> Vec<i32> {
-(1..=50).map(|n| (2 * n).pow(2)).collect()
+    (1..=50).map(|n| (2 * n as i32).pow(2)).collect()
 }
